Created a BaseGraph class so that the graph algorithm  implementations will not be cluttered with copied code
